LPD-45406 move to utils
